Re-worked bot to use cogs and discordUI slash_options seem to be broken--need to fix
